+++ scuba/__main__.py
@@ -125,6 +125,9 @@ class ScubaDive(object):
             writeln(s, '      {0} => {1} {2}'.format(hostpath, contpath, options))
 
         writeln(s, '   user_command: {0}'.format(self.user_command))
+        writeln(s, '   context:')
+        writeln(s, '     script: ' + str(self.context.script)) 
+        writeln(s, '     image:  ' + str(self.context.image)) 
 
         return s.getvalue()
 
@@ -231,7 +234,7 @@ class ScubaDive(object):
 
         # Process any aliases
         try:
-            script = self.config.process_command(self.user_command)
+            context = self.config.process_command(self.user_command)
         except ConfigError as cfgerr:
             raise ScubaError(str(cfgerr))
 
@@ -240,14 +243,14 @@ class ScubaDive(object):
         default CMD is run. Because we set the entrypiont, scuba must emulate the
         default behavior itself.
         '''
-        if len(script) == 0:
+        if not context.script:
             # No user-provided command; we want to run the image's default command
             verbose_msg('No user command; getting command from image')
             try:
-                script = [get_image_command(self.config.image)]
+                context.script = [get_image_command(context.image)]
             except DockerError as e:
                 raise ScubaError(str(e))
-            verbose_msg('{0} Cmd: "{1}"'.format(self.config.image, script[0]))
+            verbose_msg('{0} Cmd: "{1}"'.format(context.image, context.script[0]))
 
         # The user command is executed via a generated shell script
         with self.open_scubadir_file('command.sh', 'wt') as f:
@@ -255,8 +258,10 @@ class ScubaDive(object):
             writeln(f, '#!/bin/sh')
             writeln(f, '# Auto-generated from scuba')
             writeln(f, 'set -e')
-            for cmd in script:
+            for cmd in context.script:
                 writeln(f, shell_quote_cmd(cmd))
+
+        self.context = context
 
 
 
@@ -329,7 +334,7 @@ class ScubaDive(object):
         args += self.options
 
         # Docker image
-        args.append(self.config.image)
+        args.append(self.context.image)
 
         # Command to run in container
         args += self.docker_cmd