+++ tests/test_config.py
@@ -115,8 +115,8 @@ class TestConfig(TestCase):
         config = scuba.config.load_config('.scuba.yml')
         assert_equals(config.image, 'busybox')
         assert_equals(len(config.aliases), 2)
-        assert_seq_equal(config.aliases['foo'], [['bar']])
-        assert_seq_equal(config.aliases['snap'], [['crackle', 'pop']])
+        assert_seq_equal(config.aliases['foo'].script, [['bar']])
+        assert_seq_equal(config.aliases['snap'].script, [['crackle', 'pop']])
 
 
 
@@ -188,7 +188,7 @@ class TestConfig(TestCase):
                 image = 'na',
                 )
         result = cfg.process_command([])
-        assert_equal(result, [])
+        assert_equal(result.script, None)
 
 
     def test_process_command_no_aliases(self):
@@ -197,7 +197,7 @@ class TestConfig(TestCase):
                 image = 'na',
                 )
         result = cfg.process_command(['cmd', 'arg1', 'arg2'])
-        assert_equal(result, [['cmd', 'arg1', 'arg2']])
+        assert_equal(result.script, [['cmd', 'arg1', 'arg2']])
 
     def test_process_command_aliases_unused(self):
         '''process_command handles unused aliases'''
@@ -209,7 +209,7 @@ class TestConfig(TestCase):
                     ),
                 )
         result = cfg.process_command(['cmd', 'arg1', 'arg2'])
-        assert_equal(result, [['cmd', 'arg1', 'arg2']])
+        assert_equal(result.script, [['cmd', 'arg1', 'arg2']])
 
     def test_process_command_aliases_used_noargs(self):
         '''process_command handles aliases with no args'''
@@ -221,7 +221,7 @@ class TestConfig(TestCase):
                     ),
                 )
         result = cfg.process_command(['apple', 'arg1', 'arg2'])
-        assert_equal(result, [['banana', 'arg1', 'arg2']])
+        assert_equal(result.script, [['banana', 'arg1', 'arg2']])
 
     def test_process_command_aliases_used_withargs(self):
         '''process_command handles aliases with args'''
@@ -233,7 +233,7 @@ class TestConfig(TestCase):
                     ),
                 )
         result = cfg.process_command(['apple', 'arg1', 'arg2'])
-        assert_equal(result, [['banana', 'cherry', 'pie is good', 'arg1', 'arg2']])
+        assert_equal(result.script, [['banana', 'cherry', 'pie is good', 'arg1', 'arg2']])
 
     def test_process_command_multiline_aliases_used(self):
         '''process_command handles multiline aliases'''
@@ -248,7 +248,7 @@ class TestConfig(TestCase):
                     ),
                 )
         result = cfg.process_command(['apple'])
-        assert_equal(result, [
+        assert_equal(result.script, [
             ['banana', 'cherry', 'pie is good'],
             ['so', 'is', 'peach'],
         ])
@@ -266,6 +266,23 @@ class TestConfig(TestCase):
                     ),
                 )
         assert_raises(scuba.config.ConfigError, cfg.process_command, ['apple', 'ARGS', 'NOT ALLOWED'])
+
+    def test_process_command_alias_overrides_image(self):
+        '''aliases can override the image'''
+        cfg = scuba.config.ScubaConfig(
+                image = 'default',
+                aliases = dict(
+                    apple = dict(
+                        script = [
+                            'banana cherry "pie is good"',
+                            'so is peach',
+                        ],
+                        image = 'overridden',
+                    ),
+                ),
+            )
+        result = cfg.process_command(['apple'])
+        assert_equal(result.image, 'overridden')
 
 
     ############################################################################