+++ CHANGELOG.md
@@ -5,6 +5,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 ### Added
 - Added support for enhanced aliases (#67)
+- Added support for per-alias image specification (#68)
 
 ### Changed
 - All ancillary files are bind-mounted via single temp dir